+++ src/test/java/racingcar/model/GameTurnTest.java
@@ -1,10 +1,8 @@
-package racingcar.modeltest;
+package racingcar.model;
 
 import static org.assertj.core.api.Assertions.*;
 
 import org.junit.jupiter.api.Test;
-
-import racingcar.model.GameTurn;
 
 public class GameTurnTest {
 	@Test
